Refactoring SOLR SearchService
- avoiding instance creation at each API request
- extracting commons interfaces
- service definition
