+++ .gitignore
@@ -1,8 +1,8 @@
 tmp
 *.pyc
 *.iws
+*.log
 test-result
-chromedriver.log
 .DS_Store
 .idea
 logs
