+++ web-core/app/core/indexing/IndexingService.scala
@@ -161,23 +161,5 @@ object IndexingService extends SolrServer {
     commit
   }
 
-  def deleteOrphansBySpec(orgId: String, spec: String, startIndexing: DateTime)(implicit configuration: OrganizationConfiguration) {
-    val fmt = DateTimeFormat.forPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
-    val deleteQuery = SPEC.tag + ":" + spec + " AND " + ORG_ID.key + ":" + orgId + " AND timestamp:[* TO " + fmt.print(startIndexing.minusSeconds(15)) + "]"
-    val orphans = getSolrServer(configuration).query(new SolrQuery(deleteQuery)).getResults.getNumFound
-    if (orphans > 0) {
-      try {
-        val deleteResponse = getStreamingUpdateServer(configuration).deleteByQuery(deleteQuery)
-        deleteResponse.getStatus
-        commit
-        log.info("Deleting orphans %s from dataset from Solr Index: %s".format(orphans.toString, deleteQuery))
-      } catch {
-        case e: Exception => Logger.info("Unable to remove orphans for %s because of %s".format(spec, e.getMessage))
-      }
-    } else
-      log.info("No orphans found for dataset in Solr Index: %s".format(deleteQuery))
-
-  }
-
 }
